Add missed namespace renames.
