+++ web/app/Services/TagEnforcement/TagDecision.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Wikijump\Services\TagEngine;
+namespace Wikijump\Services\TagEnforcement;
 
 /**
  * The TagDecision Class stores the validation information for the TagEngine to return, including whether the inputted tags are valid, and if not valid, the reason for the invalidity.
